an initial decomposition of wasm function
