+++ vm/src/main/scala/fluence/vm/AsmbleWasmVm.scala
@@ -47,26 +47,15 @@ import scala.util.Try
  *                               that was previously allocated by allocateFunction
  */
 class AsmbleWasmVm(
-  private val functionsIndex: WasmFnIndex,
   private val modules: WasmModules,
   private val hasher: Hasher[Array[Byte], Array[Byte]],
-  private val allocateFunctionName: Option[WasmFunction],
-  private val deallocateFunctionName: Option[WasmFunction]
 ) extends WasmVm {
-
-  // TODO: now it is assumed that allocation/deallocation functions placed together in the first module.
-  // In the future it has to be refactored.
-  private val allocateFunction: Option[WasmFunction] =
-    functionsIndex.get(FunctionId(modules.head.name, AsmExtKt.getJavaIdent(allocateFunctionName)))
-
-  private val deallocateFunction: Option[WasmFunction] =
-    functionsIndex.get(FunctionId(modules.head.name, AsmExtKt.getJavaIdent(deallocateFunctionName)))
 
   override def invoke[F[_]: LiftIO: Monad](
     moduleName: Option[String],
     fnArgument: Array[Byte]
   ): EitherT[F, InvokeError, Option[Array[Byte]]] = {
-    val functionId = FunctionId(moduleName, AsmExtKt.getJavaIdent(fnName))
+    val functionId = FunctionId(moduleName, AsmExtKt.getJavaIdent("invoke"))
 
     for {
       // Finds java method(Wasm function) in the index by function id
@@ -116,39 +105,6 @@ class AsmbleWasmVm(
           } yield resultHash
       }
       .map(ByteVector(_))
-
-  // TODO : In the future, it should be rewritten with cats.effect.resource
-  /**
-   * Allocates memory in Wasm module of supplied size by allocateFunction.
-   *
-   * @param size size of memory that need to be allocated
-   * @tparam F a monad with an ability to absorb 'IO'
-   */
-  private def allocate[F[_]: LiftIO: Monad](size: Int): EitherT[F, InvokeError, AnyRef] = {
-    allocateFunction match {
-      case Some(fn) => fn(size.asInstanceOf[AnyRef] :: Nil)
-      case _ =>
-        EitherT.leftT(
-          NoSuchFnError(s"Unable to find the function for memory allocation with the name=$allocateFunctionName")
-        )
-    }
-  }
-
-  /**
-   * Deallocates previously allocated memory in Wasm module by deallocateFunction.
-   *
-   * @param offset address of memory to deallocate
-   * @tparam F a monad with an ability to absorb 'IO'
-   */
-  private def deallocate[F[_]: LiftIO: Monad](offset: Int, size: Int): EitherT[F, InvokeError, AnyRef] = {
-    deallocateFunction match {
-      case Some(fn) => fn(offset.asInstanceOf[AnyRef] :: size.asInstanceOf[AnyRef] :: Nil)
-      case _ =>
-        EitherT.leftT(
-          NoSuchFnError(s"Unable to find the function for memory deallocation with the name=$deallocateFunctionName")
-        )
-    }
-  }
 
   /**
    * Preprocesses a Wasm function argument: injects it into Wasm module memory (through injectArrayIntoWasmModule)
@@ -284,34 +240,6 @@ object AsmbleWasmVm {
   case class FunctionId(moduleName: Option[String], functionName: String) {
     override def toString =
       s"'${ModuleInstance.nameAsStr(moduleName)}.$functionName'"
-  }
-
-  /**
-   * Representation for each Wasm function. Contains reference to module instance
-   * and java method [[java.lang.reflect.Method]].
-   *
-   * @param javaMethod a java method [[java.lang.reflect.Method]] for calling function.
-   * @param module the object the underlying method is invoked from.
-   *               This is an instance for the current module, it contains
-   *               all inner state of the module, like memory.
-   */
-  case class WasmFunction(
-    fnId: FunctionId,
-    javaMethod: Method,
-    module: ModuleInstance
-  ) {
-
-    /**
-     * Invokes this function with arguments.
-     *
-     * @param args arguments for calling this function.
-     * @tparam F a monad with an ability to absorb 'IO'
-     */
-    def apply[F[_]: Functor: LiftIO](args: List[AnyRef]): EitherT[F, InvokeError, AnyRef] =
-      EitherT(IO(javaMethod.invoke(module.instance, args: _*)).attempt.to[F])
-        .leftMap(e ⇒ TrapError(s"Function $this with args: $args was failed", Some(e)))
-
-    override def toString: String = fnId.toString
   }
 
 }
